Clarify national literacy tooltips
Closes #216.

The stock tooltip for national literacy is unfortunate because it
displays average literacy within *states* on the one hand, but displays
overall *national* literacy gain (i.e. also including colonies).

With this change the difference in nature between the two quantities is
made explicit. In addition the research tooltip is amended to also make
an explicit reference to average *state* literacy, thus preserving the
original connection between the two tooltips.

The wording of the literacy tooltip is flawed, but it’s hard to do
better due to game limitations. In particular the second sentence is
repurposed in other tooltips.