Follow-up for the LUN/LUA core fixes
This is a follow-up to:

> * Fixes for LUN/LUA core removals in Congo.

which was a part of 8eb31840d3b738d690de61bc316cea16279db383.

Just like it, this corrects `LUA` (Luang Prabang) core removals to `LUN`
(Lunda) when they concern the Congo/Zambia part of the world. A review
of all `LUA` occurrences suggest that all remaining mentions really do
concern Luang Prabang now.

Signed-off-by: Miko <65079739+mikoxyz@users.noreply.github.com>